Change: Rename module containing the EntityComponent
The module should be named after the component too.
