+++ src/web/entity/EntityComponent.jsx
