+++ src/web/pages/notes/Component.jsx
@@ -19,11 +19,10 @@ import {hasId} from 'gmp/utils/id';
 import {isDefined, isArray} from 'gmp/utils/identity';
 import {shorten} from 'gmp/utils/string';
 import React from 'react';
-import EntityComponent from 'web/entity/Component';
+import EntityComponent from 'web/entity/EntityComponent';
 import NoteDialog from 'web/pages/notes/Dialog';
 import PropTypes from 'web/utils/PropTypes';
 import withGmp from 'web/utils/withGmp';
-
 
 class NoteComponent extends React.Component {
   constructor(...args) {
