+++ src/web/pages/nvts/Component.jsx
@@ -4,7 +4,7 @@
  */
 
 import React from 'react';
-import EntityComponent from 'web/entity/Component';
+import EntityComponent from 'web/entity/EntityComponent';
 import NoteComponent from 'web/pages/notes/Component';
 import OverrideComponent from 'web/pages/overrides/Component';
 import PropTypes from 'web/utils/PropTypes';
