+++ src/web/pages/reportconfigs/Component.jsx
@@ -6,11 +6,10 @@
 import _ from 'gmp/locale';
 import {isDefined} from 'gmp/utils/identity';
 import React from 'react';
-import EntityComponent from 'web/entity/Component';
+import EntityComponent from 'web/entity/EntityComponent';
 import ReportConfigDialog from 'web/pages/reportconfigs/Dialog';
 import PropTypes from 'web/utils/PropTypes';
 import withGmp from 'web/utils/withGmp';
-
 
 class ReportConfigComponent extends React.Component {
   constructor(...args) {
