+++ src/web/pages/targets/Component.jsx
@@ -8,14 +8,13 @@ import {YES_VALUE} from 'gmp/parser';
 import {first} from 'gmp/utils/array';
 import {isDefined} from 'gmp/utils/identity';
 import React from 'react';
-import EntityComponent from 'web/entity/Component';
+import EntityComponent from 'web/entity/EntityComponent';
 import CredentialsDialog from 'web/pages/credentials/Dialog';
 import PortListDialog from 'web/pages/portlists/Dialog';
 import TargetDialog from 'web/pages/targets/Dialog';
 import PropTypes from 'web/utils/PropTypes';
 import {UNSET_VALUE} from 'web/utils/Render';
 import withGmp from 'web/utils/withGmp';
-
 
 const DEFAULT_PORT_LIST_ID = '33d0cd82-57c6-11e1-8ed1-406186ea4fc5'; // All IANA assigned TCP 2012-02-10
